pyproject.toml
[project]
dynamic = ["version"]
name = "so_campaign_manager"
description = "A campaign manager to execute SO mapmaking campaign"
readme = 'README.md'
requires-python = '>=3.11,<3.13'
authors = [{ name = 'Giannis Paraskevakos', email = 'iparask@princeton.edu' }]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "numpy",
    "pydantic>=2.0",
    "radical.pilot; sys_platform != 'darwin'",
    "networkx",
    "toml",
    "click",
    "sotodlib",
    "astral",
    "slurmise; sys_platform != 'darwin'",
    "humanfriendly",
]

[build-system]
requires = ["setuptools>=61", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
version_file = "src/socm/__about__.py"

[tool.taskipy.tasks]
check = "mypy --install-types --non-interactive src/socm tests"
test = "pytest --cov=socm --random-order"
fmt-check = "ruff format --check . && ruff check ."

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
# This installs the file with the package
"socm" = ["configs/*.toml"]

[project.scripts]
socm = "socm.__main__:main"

[tool.coverage.run]
source_pkgs = ["socm", "tests"]
branch = true
parallel = true
omit = ["src/socm/_version.py"]

[tool.coverage.paths]
socm = ["src/socm", "*/so_mapmaking_campaign_manager/src/socm"]
tests = ["tests", "*/so_mapmaking_campaign_manager/tests"]

[tool.coverage.report]
exclude_lines = ["pragma: no cover"]
exclude_also = ["if TYPE_CHECKING:"]
omit = ["*/so_mapmaking_campaign_manager/src/socm/__main__.py"]

[tool.isort]
line_length = 120
profile = "black"
skip = ["docs", ".github", ".vscode", "container", "env", "env311", "build"]

[tool.ruff]
line-length = 120
exclude = ["docs", ".github", ".vscode", "container", "env", "env311", "build"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = false

# Third-party libraries without stubs
[[tool.mypy.overrides]]
module = [
    "radical.*",
    "slurmise.*",
    "sotodlib.*",
    "astropy.*",
    "humanfriendly",
]
ignore_missing_imports = true

[dependency-groups]
dev = [
    "ruff",
    "darker",
    "flake8",
    "isort",
    "coverage",
    "pytest-cov>=2.6",
    "coveralls>=1.5",
    "pytest>=4.6",
    "hypothesis",
    "pre-commit",
    "mypy>=1.0.0",
    "taskipy",
    "types-networkx",
    "types-pytz",
    "types-toml",
    "types-mock",
    "ipython"
]
docs = [
    "sphinx",
    "sphinx_rtd_theme",
    "sphinxcontrib-napoleon",
    "sphinx-autodoc-typehints",
    "myst-parser",
]
